# Conflicts:
#	src/main/java/com/SaiGonEats/SaiGonEats/SecurityConfig.java
#	src/main/java/com/SaiGonEats/SaiGonEats/controller/HomeController.java
#	src/main/java/com/SaiGonEats/SaiGonEats/controller/MenuItemController.java
#	src/main/java/com/SaiGonEats/SaiGonEats/controller/OrderController.java
#	src/main/java/com/SaiGonEats/SaiGonEats/controller/ShoppingCartController.java
#	src/main/java/com/SaiGonEats/SaiGonEats/model/Order.java
#	src/main/java/com/SaiGonEats/SaiGonEats/model/OrderDetail.java
#	src/main/java/com/SaiGonEats/SaiGonEats/repository/MenuItemRepository.java
#	src/main/java/com/SaiGonEats/SaiGonEats/repository/OrderRepository.java
#	src/main/java/com/SaiGonEats/SaiGonEats/service/MenuItemService.java
#	src/main/java/com/SaiGonEats/SaiGonEats/service/OrderService.java
#	src/main/java/com/SaiGonEats/SaiGonEats/service/ShoppingCartService.java
#	src/main/resources/static/css/main.css
#	src/main/resources/static/image/18.png
#	src/main/resources/static/image/5.png
#	src/main/resources/static/image/OM2552-Light-Grey-Soft-Backdrop_2000x.webp
#	src/main/resources/static/image/Screenshot 2024-06-08 233327.png
#	src/main/resources/static/image/Screenshot 2024-06-08 233336.png
#	src/main/resources/static/image/Screenshot 2024-06-08 233719.png
#	src/main/resources/static/image/Screenshot 2024-06-09 115432.png
#	src/main/resources/static/image/Screenshot 2024-06-09 115841.png
#	src/main/resources/static/image/Screenshot 2024-06-09 120622.png
#	src/main/resources/static/image/Screenshot 2024-06-09 121226.png
#	src/main/resources/static/image/original.png
#	src/main/resources/templates/cart/view.html
#	src/main/resources/templates/home/home.html
#	src/main/resources/templates/item/detail.html
#	src/main/resources/templates/item/form.html
#	src/main/resources/templates/item/update-form.html
#	src/main/resources/templates/layout.html
#	src/main/resources/templates/menu/list.html
#	src/main/resources/templates/order/list.html
#	src/main/resources/templates/order/view.html

+++ src/main/java/com/SaiGonEats/SaiGonEats/model/Order.java
@@ -24,7 +24,7 @@ public class Order {
 
     private LocalDateTime orderDate;
 
-    private Double totalAmount;
+    private int totalAmount;
 
     private String status;
